enterprise: upgrade mysql chart
The charts.helm.sh repo has reached EOL since 2020.
https://github.com/helm/charts/blob/master/stable/mysql/README.md

Switch to the bitnami chart which has the most stars on Artifact Hub.
https://artifacthub.io/packages/search?ts_query_web=mysql&sort=stars&page=1

This chart is licensed under Apache License 2.0 so it should be safe to
use.

Note that for a production-ready deployment of MySQL, users should
consider leveraging a cloud-hosted service, such as AWS RDS or GCP
CloudSQL. MySQL also provides an official Operator (1) which could be
installed as a helm-chart.

(1): https://dev.mysql.com/doc/mysql-operator/en/mysql-operator-introduction.html